Rename FileSliceReader.h to ReaderSlice.h
Also rename associated unit test file
