+++ OP2Utility.vcxproj
@@ -183,7 +183,7 @@
     <ClInclude Include="src\Sprite\OP2BmpLoader.h" />
     <ClInclude Include="src\Sprite\PaletteHeader.h" />
     <ClInclude Include="src\Sprite\SectionHeader.h" />
-    <ClInclude Include="src\Stream\FileSliceReader.h" />
+    <ClInclude Include="src\Stream\ReaderSlice.h" />
     <ClInclude Include="src\Stream\FileReader.h" />
     <ClInclude Include="src\Stream\FileWriter.h" />
     <ClInclude Include="src\Stream\ForwardSeekableReader.h" />
