+++ OP2Utility.vcxproj.filters
@@ -81,9 +81,6 @@
     <ClInclude Include="src\Archive\WaveFile.h">
       <Filter>Archive</Filter>
     </ClInclude>
-    <ClInclude Include="src\Stream\FileSliceReader.h">
-      <Filter>Stream</Filter>
-    </ClInclude>
     <ClInclude Include="include\OP2Utility.h">
       <Filter>include</Filter>
     </ClInclude>
@@ -179,6 +176,9 @@
     </ClInclude>
     <ClInclude Include="src\Map\SavedGameUnits.h">
       <Filter>Map</Filter>
+    </ClInclude>
+    <ClInclude Include="src\Stream\ReaderSlice.h">
+      <Filter>Stream</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
