+++ src/Archive/VolFile.cpp
@@ -1,5 +1,5 @@
 #include "VolFile.h"
-#include "../Stream/FileSliceReader.h"
+#include "../Stream/ReaderSlice.h"
 #include "../XFile.h"
 #include <stdexcept>
 #include <algorithm>
