+++ test/OP2UtilityTest.vcxproj
@@ -58,7 +58,7 @@
     <ClCompile Include="Sprite\ArtReader.test.cpp" />
     <ClCompile Include="Sprite\ArtWriter.test.cpp" />
     <ClCompile Include="Stream\FileReader.test.cpp" />
-    <ClCompile Include="Stream\FileSliceReader.test.cpp" />
+    <ClCompile Include="Stream\ReaderSlice.test.cpp" />
     <ClCompile Include="Stream\FileWriter.test.cpp" />
     <ClCompile Include="Stream\MemoryStreamReader.test.cpp" />
     <ClCompile Include="XFile.test.cpp" />
