+++ test/OP2UtilityTest.vcxproj.filters
@@ -7,9 +7,6 @@
     <ClCompile Include="Stream\FileReader.test.cpp">
       <Filter>Stream</Filter>
     </ClCompile>
-    <ClCompile Include="Stream\FileSliceReader.test.cpp">
-      <Filter>Stream</Filter>
-    </ClCompile>
     <ClCompile Include="Map\MapReader.test.cpp">
       <Filter>Map</Filter>
     </ClCompile>
@@ -57,6 +54,9 @@
     </ClCompile>
     <ClCompile Include="Map\SavedGameUnits.test.cpp">
       <Filter>Map</Filter>
+    </ClCompile>
+    <ClCompile Include="Stream\ReaderSlice.test.cpp">
+      <Filter>Stream</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
